refactor: modularize front-end logic
